fix(install): Installation prompts for omz and paru typo

+++ install_packages.sh
@@ -58,42 +58,53 @@ if [[ "$ohmyzsh_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
 
   # Install Oh My Zsh plugins
   ## ZSH Syntax Highlighting
-  git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZDOTDIR:-$HOME}/omz/plugins/zsh-syntax-highlighting
+  read -p "Do you want to install the ZSH Syntax Highlighting plugin? [y/N] " syntax_highlighting_response
+  if [[ "$syntax_highlighting_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+    git clone https://github.com/zsh-users/zsh-syntax-highlighting.git ${ZDOTDIR:-$HOME}/.config/omz/plugins/zsh-syntax-highlighting
+  fi
 
   ## ZSH Autosuggestions
-  git clone https://github.com/zsh-users/zsh-autosuggestions ${ZDOTDIR:-$HOME}/omz/plugins/zsh-autosuggestions
+  read -p "Do you want to install the ZSH Autosuggestions plugin? [y/N] " autosuggestions_response
+  if [[ "$autosuggestions_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+    git clone https://github.com/zsh-users/zsh-autosuggestions ${ZDOTDIR:-$HOME}/.config/omz/plugins/zsh-autosuggestions
+  fi
 
   ## ZSH Autocomplete
-  git clone https://github.com/marlonrichert/zsh-autocomplete ${ZDOTDIR:-$HOME}/omz/plugins/zsh-autocomplete
+  read -p "Do you want to install the ZSH Autocomplete plugin? [y/N] " autocomplete_response
+  if [[ "$autocomplete_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+    git clone https://github.com/marlonrichert/zsh-autocomplete ${ZDOTDIR:-$HOME}/.config/omz/plugins/zsh-autocomplete
+  fi
 fi
 
 # Install paru
-read -p "Do you want to install Paru? [y/N] " paru_response
-if [[ "$paru_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
-  git clone https://aur.archlinux.org/paru.git
-  cd paru
-  makepkg -si
-  cd ..
-  rm -rf paru
-fi
+if [[ "$OS" == "arch" ]]; then
+  read -p "Do you want to install Paru? [y/N] " paru_response
+  if [[ "$paru_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+    git clone https://aur.archlinux.org/paru.git
+    cd paru
+    makepkg -si
+    cd ..
+    rm -rf paru
+  fi
 
-# Using paru
-## Install packages from the AUR
-read -p "Do you want to install Discord? [y/N] " discord_response
-if [[ "$discord_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
-  paru -S discord
-fi
+  # Using paru
+  ## Install packages from the AUR
+  read -p "Do you want to install Discord (AUR)? [y/N] " discord_response
+  if [[ "$discord_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+    paru -S discord
+  fi
 
-read -p "Do you want to install Google Chrome? [y/N] " chrome_response
-if [[ "$chrome_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
-  paru -S google-chrome
-fi
+  read -p "Do you want to install Google Chrome (AUR)? [y/N] " chrome_response
+  if [[ "$chrome_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+    paru -S google-chrome
+  fi
 
-read -p "Do you want to install Snapd? [y/N] " snapd_response
-if [[ "$snapd_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
-  paru -S snapd
-  sudo systemctl enable --now snapd.socket
-  sudo ln -s /var/lib/snapd/snap /snap
+  read -p "Do you want to install Snapd (AUR)? [y/N] " snapd_response
+  if [[ "$snapd_response" =~ ^([yY][eE][sS]|[yY])$ ]]; then
+    paru -S snapd
+    sudo systemctl enable --now snapd.socket
+    sudo ln -s /var/lib/snapd/snap /snap
+  fi
 fi
 
 printf "\n\n Packages installed successfully! ヽ(・∀・)ﾉ\n\n"
